admin-service/requirements.txt
django==3.2
psycopg2==2.9.4
python-dotenv==0.21.0
django-cors-headers==3.13.0
django-extensions==3.2.1
django-split-settings==1.2.0
django-debug-toolbar==3.4.0
uwsgi==2.0.21
tabulate==0.9.0
sentry-sdk[django]==1.15.0
python_logstash==0.4.8